01_GCP_Native/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.12.1
cloudevents==1.11.0
colorama==0.4.6
cryptography==43.0.1
Deprecated==1.2.14
deprecation==2.1.0
Flask==3.0.3
frozenlist==1.4.1
functions-framework==3.8.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.25.2
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==2.1.5
multidict==6.1.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.4
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.13.1
zipp==3.20.2
